requirements.txt
streamlit>=1.22.0
pandas>=1.5.0
python-dotenv>=1.0.0
requests>=2.28.0 